feat: buttons change gameTable view

+++ src/components/landingPage/GamesTableItem.tsx
@@ -2,7 +2,7 @@ import { IGame } from "../../models/IGame";
 import "../../pages/LandingPage.css";
 import { CgTrash } from "react-icons/cg"
 import keycloak from "../../keycloak";
-import {removeGame} from "../../store/slices/gamesSlice"
+import { removeGame } from "../../store/slices/gamesSlice"
 import { useAppDispatch } from "../../store/hooks";
 import { useNavigate } from "react-router-dom";
 
@@ -14,7 +14,12 @@ function GamesTableItem({ game }: { game: IGame }) {
         const dispatch = useAppDispatch()
 
         const deleteGame = () => {
-                dispatch(removeGame(game.id))
+                // eslint-disable-next-line no-restricted-globals
+                if (confirm("Are you sure you want to delete this game?")) {
+                        console.log("DELETED")
+                        dispatch(removeGame(game.id))
+                }
+
         }
 
         const handleClick = () => {
@@ -22,11 +27,11 @@ function GamesTableItem({ game }: { game: IGame }) {
         };
 
         return (<>
-                <tr className="ms-6" onClick={handleClick}>
-                        <td> {game.title} </td>
+                <tr className="ms-6" >
+                        <td onClick={handleClick} role={"button"}> {game.title} </td>
                         <td>22</td>
                         <td>{game.state}</td>
-                        <td>{isAdmin && <button onClick={deleteGame} className="btn-delete"><CgTrash /></button>} </td>
+                        <td>{isAdmin && <button onClick={deleteGame} className="btn-delete"><CgTrash className="bosspann" /></button>} </td>
                 </tr>
         </>)
